seat-management-backend/requirements.txt
Django==5.0
Django REST framework==3.15.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
mysqlclient==2.2.0
redis==5.0.1
requests==2.31.0
Pillow==10.2.0
